requirements.txt
Django==2.2.3
django-model-utils==3.2.0
django-mptt==0.10.0
django-taggit==1.1.0
djangorestframework==3.10.2
django-allauth==0.39.1
channels==2.2.0
# mysqlclient==1.4.2.post1
Cython==0.29.13
